orchestrator-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
crewai>=0.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
pytest
pytest-asyncio
pyyaml
python-multipart 